chore(experiments): rename RunningTimeNew → RunningTime
The "New" suffix on RunningTimeNew and the sibling RunningTimeCalculatorNew/
directory was scaffolding for the migration away from the old implementation.
With the old RunningTime.tsx and RunningTimeCalculator/ removed in #59002 and
#59010, the suffix is misleading — there's no "old" anymore.

- frontend/src/scenes/experiments/ExperimentView/RunningTimeNew.tsx →
  frontend/src/scenes/experiments/ExperimentView/RunningTime.tsx (renamed
  via git mv, content move; export name updated from RunningTimeNew to
  RunningTime)
- frontend/src/scenes/experiments/RunningTimeCalculatorNew/ →
  frontend/src/scenes/experiments/RunningTimeCalculator/ (directory rename
  via git mv; kea path string inside runningTimeLogic.ts updated to match)
- frontend/src/scenes/experiments/ExperimentView/Info.tsx — import and
  usage updated to the new name

Stacked on #59010 (which freed the RunningTimeCalculator/ slot).
