+++ frontend/src/scenes/experiments/ExperimentView/Info.tsx
@@ -22,7 +22,7 @@ import { modalsLogic } from '../modalsLogic'
 import { StatusTag } from './components'
 import { ExperimentDuration } from './ExperimentDuration'
 import { ExperimentReloadAction } from './ExperimentReloadAction'
-import { RunningTimeNew } from './RunningTimeNew'
+import { RunningTime } from './RunningTime'
 
 export function Info({ tabId }: Pick<ExperimentSceneLogicProps, 'tabId'>): JSX.Element {
     const {
@@ -200,7 +200,7 @@ export function Info({ tabId }: Pick<ExperimentSceneLogicProps, 'tabId'>): JSX.E
                     <div className="flex flex-col overflow-hidden items-start min-[1100px]:items-end">
                         <div className="flex flex-wrap gap-x-8 gap-y-2 justify-end">
                             {tabId && (
-                                <RunningTimeNew
+                                <RunningTime
                                     experiment={experiment}
                                     tabId={tabId}
                                     onClick={openRunningTimeConfigModal}
