+++ frontend/src/scenes/experiments/RunningTimeCalculator/calculations.test.ts
