+++ frontend/src/scenes/experiments/RunningTimeCalculator/calculations.ts
